refactor: move fingerprint output schema into homeboy-audit-contract (#8425)
Stage 2 of the homeboy-audit extraction. extension/fingerprint.rs mixed two
concerns: the fingerprint OUTPUT SCHEMA (HookRef, UnusedParam, CallSite,
DeadCodeMarker, AggregateLiteral, AggregateConstructionSeam, FingerprintOutput)
and run_fingerprint_script (the behavior that shells out to an extension's
fingerprint script to produce that output).

The 7 types are audit's data model — pure serde structs, zero deps — consumed by
the code_audit engine (~50 call sites). Move them into homeboy-audit-contract
(alongside the config schema from stage 1) and re-export them from
extension::fingerprint so every `extension::CallSite` / `extension::fingerprint::*`
path resolves unchanged. run_fingerprint_script stays in extension and imports
the types back from the contract.

This peels the fingerprint-type half of the audit<->extension coupling into the
shared contract, continuing toward extracting code_audit as homeboy-audit.

Verified: homeboy-audit-contract compiles + 16 tests pass; homeboy-core compiles
clean; `cargo check --workspace --tests` green.

Refs #8425
